DevFest25-WS/P7-LoopAgents/pyproject.toml
[project]
name = "p6-loopagents"
version = "0.1.0"
description = "Part 6: Iterative Refinement with LoopAgent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai",
    "python-dotenv",
]

